Feature/select image for detail page (#51)
* toWKT function in seperated file

* changed let and cons to var

* add comments

* 200 to 10 polygon
